BigInteger

+++ backend/src/main/java/com/tsystems/javaschool/medical/backend/component/EventStatusChanger.java
@@ -2,8 +2,10 @@ package com.tsystems.javaschool.medical.backend.component;
 
 import com.tsystems.javaschool.medical.backend.exception.EventStatusChangerExeption;
 
+import java.math.BigInteger;
+
 public interface EventStatusChanger {
 
-    void changeStatus(int eventId , String newStatus) throws EventStatusChangerExeption;
+    void changeStatus(BigInteger eventId , String newStatus) throws EventStatusChangerExeption;
 
 }
