+++ backend/src/main/java/com/tsystems/javaschool/medical/backend/dao/DrugRepository.java
@@ -3,6 +3,8 @@ package com.tsystems.javaschool.medical.backend.dao;
 import com.tsystems.javaschool.medical.backend.entities.DrugsEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DrugRepository extends CrudRepository<DrugsEntity, Integer> {
+import java.math.BigInteger;
+
+public interface DrugRepository extends CrudRepository<DrugsEntity, BigInteger> {
 
 }
