+++ backend/src/main/java/com/tsystems/javaschool/medical/backend/dao/SpecializationRepository.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import static com.tsystems.javaschool.medical.backend.util.DateUtils.getCurrentTimestamp;
@@ -50,7 +51,7 @@ public class SpecializationRepository {
     }
 
     @Transactional
-    public void delete(int id) {
+    public void delete(BigInteger id) {
         Session session = sessionFactory.getCurrentSession();
         SpecializationsEntity specializationsEntity = session.load(SpecializationsEntity.class, id);
         specializationsEntity.setDeleted(IsDeleted.Y);
