+++ backend/src/main/java/com/tsystems/javaschool/medical/backend/dto/DrugDto.java
@@ -2,19 +2,21 @@ package com.tsystems.javaschool.medical.backend.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigInteger;
+
 public class DrugDto {
     @JsonProperty("id")
-    private int id;
+    private BigInteger id;
     @JsonProperty("description")
     private String description;
     @JsonProperty("type")
     private String type;
 
-    public int getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
