+++ backend/src/main/java/com/tsystems/javaschool/medical/backend/services/ProceduresService.java
@@ -7,6 +7,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -33,7 +34,7 @@ public class ProceduresService {
         procedureRepository.create(description);
     }
 
-    public void deleteProcedure(int id) {
+    public void deleteProcedure(BigInteger id) {
         procedureRepository.delete(id);
     }
