+++ backend/src/main/java/com/tsystems/javaschool/medical/backend/services/SpecializationsService.java
@@ -7,6 +7,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -36,7 +37,7 @@ public class SpecializationsService {
         specializationRepository.create(description);
     }
 
-    public void deleteSpecialization(int id) {
+    public void deleteSpecialization(BigInteger id) {
         specializationRepository.delete(id);
     }
